Split copy tests
Split copy tests into header and source to add datapar copy tests.
