+++ libs/parallelism/algorithms/tests/unit/algorithms/copyn.cpp
@@ -1,3 +1,4 @@
+//  Copyright (c) 2021 Srinivas Yadav
 //  Copyright (c) 2014 Grant Mercer
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -5,93 +6,14 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/local/init.hpp>
-#include <hpx/modules/testing.hpp>
-#include <hpx/parallel/algorithms/copy.hpp>
 
-#include <cstddef>
 #include <iostream>
-#include <iterator>
-#include <numeric>
-#include <random>
 #include <string>
 #include <vector>
 
-#include "test_utils.hpp"
+#include "copyn_tests.hpp"
 
 ////////////////////////////////////////////////////////////////////////////
-unsigned int seed = std::random_device{}();
-std::mt19937 gen(seed);
-
-template <typename IteratorTag>
-void test_copy_n(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    hpx::copy_n(iterator(std::begin(c)), c.size(), std::begin(d));
-
-    std::size_t count = 0;
-    HPX_TEST(std::equal(std::begin(c), std::end(c), std::begin(d),
-        [&count](std::size_t v1, std::size_t v2) -> bool {
-            HPX_TEST_EQ(v1, v2);
-            ++count;
-            return v1 == v2;
-        }));
-    HPX_TEST_EQ(count, d.size());
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_copy_n(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    hpx::copy_n(policy, iterator(std::begin(c)), c.size(), std::begin(d));
-
-    std::size_t count = 0;
-    HPX_TEST(std::equal(std::begin(c), std::end(c), std::begin(d),
-        [&count](std::size_t v1, std::size_t v2) -> bool {
-            HPX_TEST_EQ(v1, v2);
-            ++count;
-            return v1 == v2;
-        }));
-    HPX_TEST_EQ(count, d.size());
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_copy_n_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    auto f = hpx::copy_n(p, iterator(std::begin(c)), c.size(), std::begin(d));
-    f.wait();
-
-    std::size_t count = 0;
-    HPX_TEST(std::equal(std::begin(c), std::end(c), std::begin(d),
-        [&count](std::size_t v1, std::size_t v2) -> bool {
-            HPX_TEST_EQ(v1, v2);
-            ++count;
-            return v1 == v2;
-        }));
-    HPX_TEST_EQ(count, d.size());
-}
-
 template <typename IteratorTag>
 void test_copy_n()
 {
@@ -111,115 +33,7 @@ void n_copy_test()
     test_copy_n<std::forward_iterator_tag>();
 }
 
-///////////////////////////////////////////////////////////////////////////////
-template <typename IteratorTag>
-void test_copy_n_exception(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::copy_n(decorated_iterator(std::begin(c),
-                        []() { throw std::runtime_error("test"); }),
-            c.size(), std::begin(d));
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<hpx::execution::sequenced_policy,
-            IteratorTag>::call(hpx::execution::seq, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_copy_n_exception(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::copy_n(policy,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            c.size(), std::begin(d));
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_copy_n_exception_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        auto f = hpx::copy_n(p,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            c.size(), std::begin(d));
-
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
+////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_copy_n_exception()
 {
@@ -243,79 +57,7 @@ void copy_n_exception_test()
     test_copy_n_exception<std::forward_iterator_tag>();
 }
 
-////////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_copy_n_bad_alloc(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_bad_alloc = false;
-    try
-    {
-        hpx::copy_n(policy,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            c.size(), std::begin(d));
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_copy_n_bad_alloc_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::vector<std::size_t> d(c.size());
-    std::iota(std::begin(c), std::end(c), gen());
-
-    bool caught_bad_alloc = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        auto f = hpx::copy_n(p,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            c.size(), std::begin(d));
-
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-    HPX_TEST(returned_from_algorithm);
-}
-
+////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_copy_n_bad_alloc()
 {
